Enhance user ID validation in Anthropic transformation
- Updated email regex pattern to improve validation accuracy
- Reintroduced user ID validation tests for both chat and completion endpoints
- Ensured valid user IDs are accepted while invalid ones (emails) are rejected
- All tests pass, confirming the robustness of the validation logic
